requirements.txt
# Core AI/ML
crewai>=0.30.0
anthropic>=0.39.0
openai>=1.0.0

# Educational Animation
manim>=0.18.0

# Audio/TTS
aiohttp>=3.9.0
pydub>=0.25.1

# Computer Vision & OCR
opencv-python>=4.8.0
pytesseract>=0.3.10
pillow>=10.0.0

# PDF Processing
PyPDF2>=3.0.0
pdf2image>=1.16.3

# Video Processing
moviepy>=1.0.3
ffmpeg-python>=0.2.0

# Web Interface
gradio>=4.19.0
streamlit>=1.29.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
asyncio
typing-extensions>=4.8.0
requests>=2.31.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0